escape to close the menu logic moved into Menu component
