+++ src/providers/MenuStateProvider.jsx
@@ -1,9 +1,6 @@
-import { createContext, useContext, useEffect, useState } from "react";
+import { createContext, useState } from "react";
 
 export const menuStateContext = createContext();
-
-// contexts
-import { refContext } from "./RefProvider";
 
 const MenuStateProvider = ({ children }) => {
   const [menuState, setMenuState] = useState(false);
@@ -13,24 +10,6 @@ const MenuStateProvider = ({ children }) => {
   const closeTheMenu = () => {
     setMenuState(false);
   };
-
-  // close the menu when escape pressed (for accessibility)
-  const { helpers: { focusOnMenuToggler } } = useContext(refContext);
-  useEffect(() => {
-    const closeMenuHandler = (e) => {
-      if (e.key === 'Escape') {
-        focusOnMenuToggler();
-        closeTheMenu(); // focusing on menu toggler does not close the menu, even if it trigges blur
-      };
-    };
-
-    // if closed, remove the event listener; if opened, add the event listener
-    if (menuState) {
-      // keydown used instead of keyup, so that when a browser popup closed with esc, menu won't close
-      document.addEventListener('keydown', closeMenuHandler);
-      return () => document.removeEventListener('keydown', closeMenuHandler);
-    }
-  }, [menuState]);
 
   return (<menuStateContext.Provider value={{ menuState, toggleMenuState, closeTheMenu }}>
     {children}
